// models/Country.ts
export interface Country {
    id: number;
    name: string;
    languages: string;
    dialing_code: string;
    capital: string;
    cities: string;
    country_detail: string;
    climate: string;
    currency: string;
    timezone: string;
    additional_info: string;
    website: string;
    is_popular: boolean;
    is_outsource_application_center: boolean;
    is_jurisdiction: boolean;
    flag: string;
}